Move cipher cracking test into cracker test file
